requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.1
django-allauth==0.47.0
django-private-storage==3.0
django-resized==0.3.11
idna==3.3
mysqlclient==2.1.0
oauthlib==3.1.1
Pillow==9.0.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
